Move ServerInfo to separate file. Expose new net functions through NET header
