+++ CMakeLists.txt
@@ -78,7 +78,7 @@ file(GLOB SOURCES
         CONFIG.H)
 
 if (EMCC)
-    file(GLOB TRANSPORT_SOURCES SRC/NET/TRANSPORT/WEBRTC/*.CPP)
+    file(GLOB TRANSPORT_SOURCES SRC/NET/TRANSPORT/WEBRTC/*.CPP SRC/NET/SERVERINFO.CPP)
 else()
     file(GLOB TRANSPORT_SOURCES SRC/NET/TRANSPORT/SDL/*.CPP)
 endif()
