+++ SRC/NET/NET.H
@@ -1,6 +1,10 @@
 #ifndef __NET_H__
 #define __NET_H__
 
+#ifdef TK_PORT_EMCC
+#include "SERVERINFO.H"
+#endif
+
 #define RECEIVERS 16 // Defines how many packets can be handled during a frame
 #define MAXDATASIZE 1024
 #define BROADCAST_ID 0xFF
@@ -30,6 +34,11 @@ namespace net
 {
 bool init();
 void deinit();
+
+#ifdef TK_PORT_EMCC
+int connect_to_lobby();
+ServerInfos get_server_list();
+#endif
 
 int create_server(const char *server_name, unsigned int max_players, const unsigned int port);
 int connect_server(const char*, const unsigned int port);
