+++ SRC/NET/TRANSPORT/TRANSPORT.H
@@ -2,6 +2,7 @@
 #define __TRANSPORT_H__
 
 #include "PORT.H"
+#include "NET/SERVERINFO.H"
 
 #include <functional>
 #include <stdint.h>
@@ -32,22 +33,6 @@ struct Received
     const Buffer buffer;
 };
 
-#if defined(TK_PORT_EMCC)
-struct ServerInfo
-{
-  ServerInfo(const char* serialized);
-  bool operator==(const ServerInfo& rhs) const;
-  bool operator!=(const ServerInfo& rhs) const;
-
-  std::string gameId;
-  std::string serverName;
-  uint8_t playerAmount;
-  uint8_t maxPlayers;
-};
-
-using ServerInfos = std::vector<ServerInfo>;
-#endif
-
 using DisconnectCallback = std::function<void(Identifier index)>;
 
 using Port = const uint16_t;
@@ -56,9 +41,7 @@ bool init(unsigned int max_connections, size_t max_packet_size);
 
 void deinit();
 
-#if defined(TK_PORT_EMCC)
 int connect_to_lobby();
-#endif
 
 bool is_server();
 
@@ -83,9 +66,7 @@ void accept_connections();
 
 namespace client
 {
-#if defined(TK_PORT_EMCC)
 ServerInfos get_server_list();
-#endif
 
 int connect_to_server(const char* server_address, Port port);
 }
